+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `--verbose` flag to make logs less noisy.
+
 ### Changed
 
 - Use apiextensions-application to remove CAPI dependency.